cloudformation-template/us-east-1-stable/backend/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
boto3>=1.35.0
requests>=2.31.0
pydantic>=2.5.0
psycopg2-binary==2.9.9
uv
strands-agents
strands-agents-tools
bedrock-agentcore
bedrock-agentcore-starter-toolkit
mcp